network inputs

+++ Assets/Scripts/Networking/InboundMessage.cs
@@ -1,9 +1,9 @@
 public readonly struct InboundMessage
 {
-    public readonly int SenderId;
+    public readonly ushort SenderId;
     public readonly byte[] Data;
 
-    public InboundMessage(int senderId, byte[] data)
+    public InboundMessage(ushort senderId, byte[] data)
     {
         SenderId = senderId;
         Data = data;
